requirements-animations.txt
-r requirements.txt
pygame
pyopengl
